fixed issue 642

+++ cvxpy/interface/numpy_interface/sparse_matrix_interface.py
@@ -39,7 +39,7 @@ class SparseMatrixInterface(NDArrayInterface):
         # Convert cvxopt sparse to coo matrix.
         if isinstance(value, list):
             return sp.csc_matrix(value, dtype=np.double).T
-        return sp.csc_matrix(value, dtype=np.double)
+        return sp.csc_matrix(value, dtype=value.dtype)
 
     def identity(self, size):
         """Return an identity matrix.
